Unify http(s) URL validation in Lamb\Http
is_valid_http_url() lived in the webmention module while network.php's
get_feeds() rolled its own filter_var-based check — three validation
styles for the same question. Move the helper to Lamb\Http (home of the
other transport primitives) and use it everywhere. The feeds filter is
marginally more lenient than FILTER_VALIDATE_URL on exotic-but-http
URLs; SimplePie still owns actual fetch failures.
